fix: removing ethereumj dependency from test-clients gradle file and adapt tests

+++ test-clients/src/main/java/com/hedera/services/bdd/suites/contract/hapi/ContractCallLocalSuite.java
@@ -182,10 +182,7 @@ public class ContractCallLocalSuite extends HapiApiSuite {
                                         resultWith()
                                                 .resultThruAbi(
                                                         decimalsABI,
-                                                        isLiteralResult(
-                                                                new Object[] {
-                                                                    BigInteger.valueOf(DECIMALS)
-                                                                }))));
+                                                        isLiteralResult(new Object[] {DECIMALS}))));
     }
 
     @Override
